Creating /export route to work through the API
