+++ README.md
@@ -1,4 +1,4 @@
-# Bookmarky Api v0.0.1.55
+# Bookmarky Api v0.0.1.56
 
 Self hosted open source bookmark tool
