+++ kubernetes-manifests/envs/prod/kustomization.yaml
@@ -6,7 +6,7 @@ resources:
   # - secrets.yaml
 images:
   - name: harbor.squid-ink.us/politeauthority/bookmarky-api
-    newTag: 0.0.1.55
+    newTag: 0.0.1.56
 patches:
   - target:
       kind: Deployment
