+++ kubernetes-manifests/envs/stage/kustomization.yaml
@@ -6,7 +6,7 @@ resources:
   # - secret.yaml
 images:
   - name: harbor.squid-ink.us/politeauthority/bookmarky-api
-    newTag: stage-0.0.1.55
+    newTag: stage-0.0.1.56
 patches:
   - target:
       kind: Deployment
