+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "bookmarky"
-version = "0.0.1.55"
+version = "0.0.1.56"
 
 requires-python = ">=3.11"
 dependencies = [
