+++ src/bookmarky/api/version.py
@@ -4,7 +4,7 @@
 
 """
 
-version = "0.0.1.55"
+version = "0.0.1.56"
 
 
 # End File: politeauthority/bookmarky-api/src/bookmarky/modules/version.py
